camel-k-runtime-jvm is misleading #101
