+++ camel-k-runtime-main/src/main/resources/log4j2.properties
