+++ camel-k-runtime-main/src/test/resources/conf.d/001/conf.properties
