+++ camel-k-runtime-main/src/test/resources/conf.d/002/conf.properties
