+++ camel-k-runtime-main/src/test/resources/log4j2-test.xml
