+++ camel-k-runtime-main/src/test/resources/my-resource.txt
