+++ camel-k-runtime-main/src/test/resources/r1.js
