+++ camel-k-runtime-main/src/test/resources/r2.mytype
