+++ pom.xml
@@ -187,7 +187,7 @@
         <module>camel-knative</module>
         <module>camel-k-maven-plugin</module>
         <module>camel-k-runtime-core</module>
-        <module>camel-k-runtime-jvm</module>
+        <module>camel-k-runtime-main</module>
 
         <module>camel-k-loader-groovy</module>
         <module>camel-k-loader-kotlin</module>
@@ -258,10 +258,9 @@
                 <artifactId>camel-k-runtime-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
-
             <dependency>
                 <groupId>org.apache.camel.k</groupId>
-                <artifactId>camel-k-runtime-jvm</artifactId>
+                <artifactId>camel-k-runtime-main</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -326,6 +325,12 @@
             <dependency>
                 <groupId>org.apache.camel.k</groupId>
                 <artifactId>camel-k-runtime-yaml</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.apache.camel.k</groupId>
+                <artifactId>camel-k-runtime-jvm</artifactId>
                 <version>${project.version}</version>
             </dependency>
         </dependencies>
